+++ src/domain/mod.rs
@@ -13,7 +13,7 @@ mod types;
 pub(crate) mod ui;
 
 pub use character::*;
-#[cfg(feature = "std")]
+#[cfg(any(feature = "barcodes", feature = "codes_2d"))]
 pub use codes::*;
 pub use common::chars_number;
 pub use constants::*;